// src/types/location.ts
export interface Location {
  slug: string
  city: string
  state: string
  tagline: string
  heroBanner: string
  owner: string
  phone: string
  email: string
  description: string
  features: string[]
  services: string[]
  calendlyUrl: string
  facebook: string
  instagram: string
  areas: string[]
}